coses issue #35
